Fix iconv checks when using internal lapack/blas
acx_lapack_ok is set to yes even when the lapack library is not
available. This causes "-lblas" to be appended to LIBS, which is a
problem when using the internal lapack/blas implementation because the
blas library is not built yet. And ultimately this causes the iconv
checks to fail because it cannot find blas.